Rename characterLoader.py to character_loader.py
